requirements.txt
# =============================================================================
# PaddleOCR — Text Detection & Text Recognition (No PPStructure)
# =============================================================================
#
# INSTALLATION ORDER:
#
# Step 1 — PaddlePaddle GPU (MUST be installed first, custom index):
#   pip install paddlepaddle-gpu==3.2.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu126/
#
# Step 2 — PaddleOCR (WITHOUT doc-parser extra — no PPStructure):
#   pip install paddleocr==3.3.2
#
# Step 3 — Remaining dependencies:
#   pip install -r requirements.txt
#
# =============================================================================

# PaddlePaddle base engine (install via custom index, not pip)
# paddlepaddle-gpu==3.2.0

# PaddleOCR — text detection + text recognition only
# NOTE: Do NOT use paddleocr[doc-parser] — that pulls in PPStructure dependencies
paddleocr==3.3.2

# Image processing
opencv-contrib-python==4.10.0.84
Pillow>=10.0,<12.0
numpy>=1.24.4,<2.0.0
scipy

# Polygon / spatial ops
shapely>=2.0
pyclipper>=1.4
